# HDDS-10215. Speed up some tests that use OmTestManagers