--- a/hadoop-hdds/test-utils/src/main/java/org/apache/ozone/test/NamedTestsBase.java
+++ b/hadoop-hdds/test-utils/src/main/java/org/apache/ozone/test/NamedTestsBase.java
@@ -1,0 +1,42 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.ozone.test;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.TestInfo;
+
+import java.lang.reflect.Method;
+
+/**
+ * Base class for JUnit tests.  Provides test method name, which can be used to create unique items.
+ */
+public abstract class NamedTestsBase {
+
+  private TestInfo info;
+
+  @BeforeEach
+  void storeTestInfo(TestInfo testInfo) {
+    this.info = testInfo;
+  }
+
+  protected String getTestName() {
+    // FIXME parameterized tests?
+    return info.getTestMethod().map(Method::getName).orElse("unknown");
+  }
+
+}