--- a/hadoop-ozone/ozone-manager/src/test/java/org/apache/hadoop/ozone/om/TestBucketManagerImpl.java
+++ b/hadoop-ozone/ozone-manager/src/test/java/org/apache/hadoop/ozone/om/TestBucketManagerImpl.java
@@ -18,7 +18,6 @@ package org.apache.hadoop.ozone.om;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.Collections;
 
@@ -43,9 +42,11 @@ import org.apache.hadoop.ozone.om.helpers.OpenKeySession;
 import org.apache.hadoop.ozone.om.helpers.BucketLayout;
 import org.apache.hadoop.ozone.om.protocol.OzoneManagerProtocol;
 import org.apache.hadoop.ozone.om.request.OMRequestTestUtils;
-import org.apache.hadoop.security.authentication.client.AuthenticationException;
-import org.junit.jupiter.api.AfterEach;
+import org.apache.ozone.test.NamedTestsBase;
+import org.junit.jupiter.api.AfterAll;
+import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.api.io.TempDir;
 
@@ -66,40 +67,36 @@ import static org.mockito.Mockito.when;
  * This class tests the Bucket Manager Implementation using Mockito.
  */
 @ExtendWith(MockitoExtension.class)
-public class TestBucketManagerImpl {
-
-  @TempDir
-  private Path folder;
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+class TestBucketManagerImpl extends NamedTestsBase {
 
   private OmTestManagers omTestManagers;
   private OzoneManagerProtocol writeClient;
 
-  @AfterEach
-  public void cleanup() throws Exception {
-    OzoneManager om = omTestManagers.getOzoneManager();
-    om.stop();
-  }
-
-  private OzoneConfiguration createNewTestPath() throws IOException {
+  @BeforeAll
+  void setup(@TempDir File folder) throws Exception {
     OzoneConfiguration conf = new OzoneConfiguration();
-    File newFolder = folder.toFile();
-    if (!newFolder.exists()) {
-      assertTrue(newFolder.mkdirs());
-    }
-    ServerUtils.setOzoneMetaDirPath(conf, newFolder.toString());
-    return conf;
-  }
+    ServerUtils.setOzoneMetaDirPath(conf, folder.toString());
 
-  private void createSampleVol() throws IOException, AuthenticationException {
-    OzoneConfiguration conf = createNewTestPath();
     omTestManagers = new OmTestManagers(conf);
     writeClient = omTestManagers.getWriteClient();
+  }
 
+  @AfterAll
+  void cleanup() throws Exception {
+    omTestManagers.getOzoneManager().stop();
+  }
+
+  public String volumeName() {
+    return getTestName().toLowerCase();
+  }
+
+  private void createSampleVol(String volume) throws IOException {
     // This is a simple hack for testing, we just test if the volume via a
     // null check, do not parse the value part. So just write some dummy value.
     OmVolumeArgs args =
         OmVolumeArgs.newBuilder()
-            .setVolume("sample-vol")
+            .setVolume(volume)
             .setAdminName("bilbo")
             .setOwnerName("bilbo")
             .build();
@@ -107,25 +104,20 @@ public class TestBucketManagerImpl {
   }
 
   @Test
-  public void testCreateBucketWithoutVolume() throws Exception {
-    OzoneConfiguration conf = createNewTestPath();
-    omTestManagers = new OmTestManagers(conf);
-    OMException omEx = assertThrows(OMException.class, () -> {
-      writeClient = omTestManagers.getWriteClient();
-
-      OmBucketInfo bucketInfo = OmBucketInfo.newBuilder()
-          .setVolumeName("sample-vol")
-          .setBucketName("bucket-one")
-          .build();
-      writeClient.createBucket(bucketInfo);
-    });
+  void testCreateBucketWithoutVolume() {
+    OmBucketInfo bucketInfo = OmBucketInfo.newBuilder()
+        .setVolumeName(volumeName())
+        .setBucketName("bucket-one")
+        .build();
+    OMException omEx = assertThrows(OMException.class, () -> writeClient.createBucket(bucketInfo));
     assertEquals(ResultCodes.VOLUME_NOT_FOUND, omEx.getResult());
     assertEquals("Volume doesn't exist", omEx.getMessage());
   }
 
   @Test
-  public void testCreateEncryptedBucket() throws Exception {
-    createSampleVol();
+  void testCreateEncryptedBucket() throws Exception {
+    String volume = volumeName();
+    createSampleVol(volume);
     KeyProviderCryptoExtension kmsProvider = omTestManagers.kmsProviderInit();
 
     String testBekName = "key1";
@@ -138,17 +130,16 @@ public class TestBucketManagerImpl {
     BucketManager bucketManager = omTestManagers.getBucketManager();
 
     OmBucketInfo bucketInfo = OmBucketInfo.newBuilder()
-        .setVolumeName("sample-vol")
+        .setVolumeName(volume)
         .setBucketName("bucket-one")
         .setBucketEncryptionKey(new
             BucketEncryptionKeyInfo.Builder().setKeyName("key1").build())
         .build();
     writeClient.createBucket(bucketInfo);
-    assertNotNull(bucketManager.getBucketInfo("sample-vol",
-        "bucket-one"));
+    assertNotNull(bucketManager.getBucketInfo(volume, "bucket-one"));
 
     OmBucketInfo bucketInfoRead =
-        bucketManager.getBucketInfo("sample-vol", "bucket-one");
+        bucketManager.getBucketInfo(volume, "bucket-one");
 
     assertEquals(bucketInfoRead.getEncryptionKeyInfo().getKeyName(),
         bucketInfo.getEncryptionKeyInfo().getKeyName());
@@ -157,67 +148,62 @@ public class TestBucketManagerImpl {
 
   @Test
   public void testCreateBucket() throws Exception {
-    createSampleVol();
+    String volume = volumeName();
+    createSampleVol(volume);
 
     BucketManager bucketManager = omTestManagers.getBucketManager();
     OmBucketInfo bucketInfo = OmBucketInfo.newBuilder()
-        .setVolumeName("sample-vol")
+        .setVolumeName(volume)
         .setBucketName("bucket-one")
         .build();
     writeClient.createBucket(bucketInfo);
-    assertNotNull(bucketManager.getBucketInfo("sample-vol",
-        "bucket-one"));
+    assertNotNull(bucketManager.getBucketInfo(volume, "bucket-one"));
   }
 
   @Test
   public void testCreateAlreadyExistingBucket() throws Exception {
-    createSampleVol();
+    String volume = volumeName();
+    createSampleVol(volume);
 
-    OMException omEx = assertThrows(OMException.class, () -> {
-      OmBucketInfo bucketInfo = OmBucketInfo.newBuilder()
-          .setVolumeName("sample-vol")
-          .setBucketName("bucket-one")
-          .build();
-      writeClient.createBucket(bucketInfo);
-      writeClient.createBucket(bucketInfo);
-    });
-    assertEquals(ResultCodes.BUCKET_ALREADY_EXISTS,
-          omEx.getResult());
+    OmBucketInfo bucketInfo = OmBucketInfo.newBuilder()
+        .setVolumeName(volume)
+        .setBucketName("bucket-one")
+        .build();
+    writeClient.createBucket(bucketInfo);
+
+    OMException omEx = assertThrows(OMException.class,
+        () -> writeClient.createBucket(bucketInfo));
+    assertEquals(ResultCodes.BUCKET_ALREADY_EXISTS, omEx.getResult());
     assertEquals("Bucket already exist", omEx.getMessage());
   }
 
   @Test
   public void testGetBucketInfoForInvalidBucket() throws Exception {
-    createSampleVol();
-    OMException exception = assertThrows(OMException.class, () -> {
-      BucketManager bucketManager = omTestManagers.getBucketManager();
-      bucketManager.getBucketInfo("sample-vol", "bucket-one");
-    });
+    String volume = volumeName();
+    createSampleVol(volume);
+
+    BucketManager bucketManager = omTestManagers.getBucketManager();
+
+    OMException exception = assertThrows(OMException.class,
+        () -> bucketManager.getBucketInfo(volume, "bucket-one"));
     assertThat(exception.getMessage()).contains("Bucket not found");
-    assertEquals(ResultCodes.BUCKET_NOT_FOUND,
-        exception.getResult());
+    assertEquals(ResultCodes.BUCKET_NOT_FOUND, exception.getResult());
   }
 
   @Test
-  public void testGetBucketInfo() throws Exception {
-    final String volumeName = "sample-vol";
+  void testGetBucketInfo() throws Exception {
+    final String volumeName = volumeName();
     final String bucketName = "bucket-one";
-
-    OzoneConfiguration conf = createNewTestPath();
-    omTestManagers = new OmTestManagers(conf);
-    writeClient = omTestManagers.getWriteClient();
 
     OMMetadataManager metaMgr = omTestManagers.getMetadataManager();
     BucketManager bucketManager = omTestManagers.getBucketManager();
     // Check exception thrown when volume does not exist
-    try {
-      bucketManager.getBucketInfo(volumeName, bucketName);
-      fail("Should have thrown OMException");
-    } catch (OMException omEx) {
-      assertEquals(ResultCodes.VOLUME_NOT_FOUND, omEx.getResult(),
-          "getBucketInfo() should have thrown " +
-              "VOLUME_NOT_FOUND as the parent volume is not created!");
-    }
+    OMException omEx = assertThrows(OMException.class,
+        () -> bucketManager.getBucketInfo(volumeName, bucketName));
+    assertEquals(ResultCodes.VOLUME_NOT_FOUND, omEx.getResult(),
+        "getBucketInfo() should have thrown " +
+            "VOLUME_NOT_FOUND as the parent volume is not created!");
+
     OmBucketInfo bucketInfo = OmBucketInfo.newBuilder()
         .setVolumeName(volumeName)
         .setBucketName(bucketName)
@@ -235,16 +221,12 @@ public class TestBucketManagerImpl {
     writeClient.createVolume(args);
     // Create bucket
     createBucket(metaMgr, bucketInfo);
+
     // Check exception thrown when bucket does not exist
-    try {
-      bucketManager.getBucketInfo(volumeName, "bucketNotExist");
-      fail("Should have thrown OMException");
-    } catch (OMException omEx) {
-      assertEquals(
-          ResultCodes.BUCKET_NOT_FOUND, omEx.getResult(),
-          "getBucketInfo() should have thrown BUCKET_NOT_FOUND " +
-              "as the parent volume exists but bucket doesn't!");
-    }
+    OMException e2 = assertThrows(OMException.class,
+        () -> bucketManager.getBucketInfo(volumeName, "bucketNotExist"));
+    assertEquals(ResultCodes.BUCKET_NOT_FOUND, e2.getResult());
+
     OmBucketInfo result = bucketManager.getBucketInfo(volumeName, bucketName);
     assertEquals(volumeName, result.getVolumeName());
     assertEquals(bucketName, result.getBucketName());
@@ -259,64 +241,68 @@ public class TestBucketManagerImpl {
 
   @Test
   public void testSetBucketPropertyChangeStorageType() throws Exception {
+    String volume = volumeName();
+    createSampleVol(volume);
 
-    createSampleVol();
     OMMetadataManager metaMgr = omTestManagers.getMetadataManager();
     BucketManager bucketManager = omTestManagers.getBucketManager();
     OmBucketInfo bucketInfo = OmBucketInfo.newBuilder()
-        .setVolumeName("sample-vol")
+        .setVolumeName(volume)
         .setBucketName("bucket-one")
         .setStorageType(StorageType.DISK)
         .build();
     createBucket(metaMgr, bucketInfo);
     OmBucketInfo result = bucketManager.getBucketInfo(
-        "sample-vol", "bucket-one");
+        volume, "bucket-one");
     assertEquals(StorageType.DISK,
         result.getStorageType());
     OmBucketArgs bucketArgs = OmBucketArgs.newBuilder()
-        .setVolumeName("sample-vol")
+        .setVolumeName(volume)
         .setBucketName("bucket-one")
         .setStorageType(StorageType.SSD)
         .build();
     writeClient.setBucketProperty(bucketArgs);
     OmBucketInfo updatedResult = bucketManager.getBucketInfo(
-        "sample-vol", "bucket-one");
+        volume, "bucket-one");
     assertEquals(StorageType.SSD,
         updatedResult.getStorageType());
   }
 
   @Test
   public void testSetBucketPropertyChangeVersioning() throws Exception {
-    createSampleVol();
+    String volume = volumeName();
+    createSampleVol(volume);
 
     BucketManager bucketManager = omTestManagers.getBucketManager();
     OmBucketInfo bucketInfo = OmBucketInfo.newBuilder()
-        .setVolumeName("sample-vol")
+        .setVolumeName(volume)
         .setBucketName("bucket-one")
         .setIsVersionEnabled(false)
         .build();
     writeClient.createBucket(bucketInfo);
     OmBucketInfo result = bucketManager.getBucketInfo(
-        "sample-vol", "bucket-one");
+        volume, "bucket-one");
     assertFalse(result.getIsVersionEnabled());
     OmBucketArgs bucketArgs = OmBucketArgs.newBuilder()
-        .setVolumeName("sample-vol")
+        .setVolumeName(volume)
         .setBucketName("bucket-one")
         .setIsVersionEnabled(true)
         .build();
     writeClient.setBucketProperty(bucketArgs);
     OmBucketInfo updatedResult = bucketManager.getBucketInfo(
-        "sample-vol", "bucket-one");
+        volume, "bucket-one");
     assertTrue(updatedResult.getIsVersionEnabled());
   }
 
   @Test
   public void testDeleteBucket() throws Exception {
-    createSampleVol();
+    String volume = volumeName();
+    createSampleVol(volume);
+
     BucketManager bucketManager = omTestManagers.getBucketManager();
     for (int i = 0; i < 5; i++) {
       OmBucketInfo bucketInfo = OmBucketInfo.newBuilder()
-          .setVolumeName("sample-vol")
+          .setVolumeName(volume)
           .setBucketName("bucket-" + i)
           .build();
       writeClient.createBucket(bucketInfo);
@@ -324,17 +310,17 @@ public class TestBucketManagerImpl {
     for (int i = 0; i < 5; i++) {
       assertEquals("bucket-" + i,
           bucketManager.getBucketInfo(
-              "sample-vol", "bucket-" + i).getBucketName());
+              volume, "bucket-" + i).getBucketName());
     }
     try {
-      writeClient.deleteBucket("sample-vol", "bucket-1");
+      writeClient.deleteBucket(volume, "bucket-1");
       assertNotNull(bucketManager.getBucketInfo(
-          "sample-vol", "bucket-2"));
+          volume, "bucket-2"));
     } catch (IOException ex) {
       fail(ex.getMessage());
     }
     OMException omEx = assertThrows(OMException.class, () -> {
-      bucketManager.getBucketInfo("sample-vol", "bucket-1");
+      bucketManager.getBucketInfo(volume, "bucket-1");
     });
     assertEquals(ResultCodes.BUCKET_NOT_FOUND,
           omEx.getResult());
@@ -343,15 +329,17 @@ public class TestBucketManagerImpl {
 
   @Test
   public void testDeleteNonEmptyBucket() throws Exception {
-    createSampleVol();
+    String volume = volumeName();
+    createSampleVol(volume);
+
     OmBucketInfo bucketInfo = OmBucketInfo.newBuilder()
-        .setVolumeName("sample-vol")
+        .setVolumeName(volume)
         .setBucketName("bucket-one")
         .build();
     writeClient.createBucket(bucketInfo);
     //Create keys in bucket
     OmKeyArgs args1 = new OmKeyArgs.Builder()
-            .setVolumeName("sample-vol")
+            .setVolumeName(volume)
             .setBucketName("bucket-one")
             .setKeyName("key-one")
             .setAcls(Collections.emptyList())
@@ -364,7 +352,7 @@ public class TestBucketManagerImpl {
     writeClient.commitKey(args1, session1.getId());
 
     OmKeyArgs args2 = new OmKeyArgs.Builder()
-            .setVolumeName("sample-vol")
+            .setVolumeName(volume)
             .setBucketName("bucket-one")
             .setKeyName("key-two")
             .setAcls(Collections.emptyList())
@@ -376,7 +364,7 @@ public class TestBucketManagerImpl {
     OpenKeySession session2 = writeClient.openKey(args2);
     writeClient.commitKey(args2, session2.getId());
     OMException omEx = assertThrows(OMException.class, () -> {
-      writeClient.deleteBucket("sample-vol", "bucket-one");
+      writeClient.deleteBucket(volume, "bucket-one");
     });
     assertEquals(ResultCodes.BUCKET_NOT_EMPTY,
           omEx.getResult());
@@ -385,10 +373,12 @@ public class TestBucketManagerImpl {
 
   @Test
   public void testLinkedBucketResolution() throws Exception {
-    createSampleVol();
+    String volume = volumeName();
+    createSampleVol(volume);
+
     ECReplicationConfig ecConfig = new ECReplicationConfig(3, 2);
     OmBucketInfo bucketInfo = OmBucketInfo.newBuilder()
-        .setVolumeName("sample-vol")
+        .setVolumeName(volume)
         .setBucketName("bucket-one")
         .setDefaultReplicationConfig(
             new DefaultReplicationConfig(
@@ -405,23 +395,23 @@ public class TestBucketManagerImpl {
     writeClient.createBucket(bucketInfo);
 
     OmBucketInfo bucketLinkInfo = OmBucketInfo.newBuilder()
-        .setVolumeName("sample-vol")
+        .setVolumeName(volume)
         .setBucketName("link-one")
-        .setSourceVolume("sample-vol")
+        .setSourceVolume(volume)
         .setSourceBucket("bucket-one")
         .build();
     writeClient.createBucket(bucketLinkInfo);
 
     OmBucketInfo bucketLink2 = OmBucketInfo.newBuilder()
-        .setVolumeName("sample-vol")
+        .setVolumeName(volume)
         .setBucketName("link-two")
-        .setSourceVolume("sample-vol")
+        .setSourceVolume(volume)
         .setSourceBucket("link-one")
         .build();
     writeClient.createBucket(bucketLink2);
 
     OmBucketInfo storedLinkBucket =
-        writeClient.getBucketInfo("sample-vol", "link-two");
+        writeClient.getBucketInfo(volume, "link-two");
     assertNotNull(storedLinkBucket.getDefaultReplicationConfig(),
         "Replication config is not set");
     assertEquals(ecConfig,
@@ -432,12 +422,12 @@ public class TestBucketManagerImpl {
     assertEquals(
         "link-two", storedLinkBucket.getBucketName());
     assertEquals(
-        "sample-vol", storedLinkBucket.getVolumeName());
+        volume, storedLinkBucket.getVolumeName());
 
     assertEquals(
         "link-one", storedLinkBucket.getSourceBucket());
     assertEquals(
-        "sample-vol", storedLinkBucket.getSourceVolume());
+        volume, storedLinkBucket.getSourceVolume());
 
     assertEquals(
         bucketInfo.getBucketLayout(),